ClimbOperator and safer extension, working end of 3/5
