+++ src/main/java/frc/robot/Constants.java
@@ -26,6 +26,7 @@ public final class Constants {
   public static final class JoyConstants {
     public static final int kDriverJoy = 0;
     public static final int kOperatorJoy = 1;
+    public static final int kOperatorClimbJoy = 2;
     public static final double kDeadband = 0.05;
   }
