+++ src/main/java/frc/robot/Robot.java
@@ -50,7 +50,7 @@ public class Robot extends TimedRobot {
     // and running subsystem periodic() methods.  This must be called from the robot's periodic
     // block in order for anything in the Command-based framework to work.
     CommandScheduler.getInstance().run();
-    //RobotContainer.m_shuffleboard.update();
+    RobotContainer.m_shuffleboard.update();
     RobotContainer.m_drive.updateMotors();
   }
